requirements.txt
bleak
pyyaml
pybluez
numpy
pandas
rich
matplotlib
tabulate
